IChannelOptions added to allow handlers to utilize rabbit Channels and logic for processing erroring messages. Minor refactoring
