+++ SimpleRabbit.NetCore/Service/BasicRabbitService.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using Microsoft.Extensions.Options;
 using RabbitMQ.Client;
+using SimpleRabbit.NetCore.Model;
 
-namespace SimpleRabbit.NetCore
+namespace SimpleRabbit.NetCore.Service
 {
     public abstract class BasicRabbitService : IDisposable
     {
