+++ SimpleRabbit.NetCore/Service/IChannelOptions.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using RabbitMQ.Client;
+
+namespace SimpleRabbit.NetCore.Service
+{
+    public interface IChannelOptions
+    {
+        void SetChannelOptions(IModel model, Action onError);
+    }
+}
