+++ SimpleRabbit.NetCore/Service/PublishService.cs
@@ -2,8 +2,9 @@
 using System.Text;
 using Microsoft.Extensions.Options;
 using RabbitMQ.Client;
+using SimpleRabbit.NetCore.Model;
 
-namespace SimpleRabbit.NetCore
+namespace SimpleRabbit.NetCore.Service
 {
     public interface IPublishService
     {
