+++ SimpleRabbit.NetCore/Service/QueueService.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Options;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
 
-namespace SimpleRabbit.NetCore
+namespace SimpleRabbit.NetCore.Service
 {
     public interface IQueueService
     {
@@ -16,9 +16,16 @@ namespace SimpleRabbit.NetCore
 
     public class QueueService : BasicRabbitService, IQueueService
     {
+        private const int MaxRetryInterval = 120;
+
         private readonly ILogger<QueueService> _logger;
 
         private readonly Timer _timer;
+        private SubscriberConfiguration _queueServiceParams;
+
+        private IMessageHandler _handler;
+        private int _retryCount;
+
         public QueueService(IOptions<RabbitConfiguration> options, ILogger<QueueService> logger) : base(options)
         {
             _logger = logger;
@@ -34,42 +41,32 @@ namespace SimpleRabbit.NetCore
             };
         }
 
-        private const int MaxRetryInterval = 120;
         private void RestartIn(TimeSpan waitInterval)
         {
             _retryCount++;
-            var interval = waitInterval.TotalSeconds * (_autoBackOff ? _retryCount : 1) % MaxRetryInterval;
-            
+            var interval = waitInterval.TotalSeconds * (_queueServiceParams.AutoBackOff ? _retryCount : 1) % MaxRetryInterval;
+
             _timer.Interval = interval * 1000;
             _logger.LogInformation($" -> restarting connection in {interval} seconds ({_retryCount}).");
             _timer.Start();
         }
 
-        private string _queue;
-        private string _tag;
-        private IMessageHandler _handler;
-        private ushort _prefetch;
-        private int _retryInterval;
-        private bool _autoBackOff;
-        private int _retryCount;
-
         public void Start(string queue, string tag, IMessageHandler handler, ushort prefetch = 1)
         {
-            _prefetch = prefetch;
-            _queue = queue;
-            _tag = tag;
+            _queueServiceParams = new SubscriberConfiguration
+            {
+                PrefetchCount = prefetch,
+                QueueName = queue,
+                ConsumerTag = tag
+            };
             _handler = handler;
+
             Start();
         }
 
         public void Start(SubscriberConfiguration subscriberConfiguration, IMessageHandler handler)
         {
-            _prefetch = subscriberConfiguration.PrefetchCount ?? 1;
-            _queue = subscriberConfiguration.QueueName;
-            _tag = subscriberConfiguration.ConsumerTag;
-            _retryInterval = subscriberConfiguration.RetryInterval;
-            _autoBackOff = subscriberConfiguration.AutoBackOff;
-            _handler = handler;
+            _queueServiceParams = subscriberConfiguration;
             Start();
         }
 
@@ -78,19 +75,26 @@ namespace SimpleRabbit.NetCore
             Stop();
             if (_handler == null)
             {
-                throw new ArgumentNullException(nameof(_handler), $"No handler provided for {_tag}");
+                throw new ArgumentNullException(nameof(_handler), $"No handler provided for {_queueServiceParams.ConsumerTag}");
             }
             try
             {
                 var consumer = new EventingBasicConsumer(Channel);
                 consumer.Received += ReceiveEvent;
-                Channel.BasicQos(0, _prefetch, false);
-                Channel.BasicConsume(_queue, false, _tag, consumer);
+
+                if (_handler is IChannelOptions options)
+                {
+                    options.SetChannelOptions(Channel,
+                        () => RestartIn(TimeSpan.FromSeconds(_queueServiceParams.RetryInterval)));
+                }
+
+                Channel.BasicQos(0, _queueServiceParams.PrefetchCount ?? 1, false);
+                Channel.BasicConsume(_queueServiceParams.QueueName, false, _queueServiceParams.ConsumerTag, consumer);
             }
             catch (Exception e)
             {
                 _logger.LogError(e, e.Message);
-                RestartIn(TimeSpan.FromSeconds(_retryInterval));
+                RestartIn(TimeSpan.FromSeconds(_queueServiceParams.RetryInterval));
             }
         }
 
@@ -116,8 +120,8 @@ namespace SimpleRabbit.NetCore
             {
                 // error processing message
                 _logger.LogError(ex, $"{ex.Message} -> {args.DeliveryTag}: {args.BasicProperties.MessageId}");
-                channel.BasicCancel(_tag);
-                RestartIn(TimeSpan.FromSeconds(_retryInterval));
+                channel.BasicCancel(_queueServiceParams.ConsumerTag);
+                RestartIn(TimeSpan.FromSeconds(_queueServiceParams.RetryInterval));
             }
         }
     }
